requirements.txt
Pillow
RPi.GPIO
spidev
requests